Fixed issue with synthetic airspeed in ECU_Lite CAN driver.

+++ libraries/AP_EFI/AP_EFI_ECU_Lite.h
@@ -89,5 +89,5 @@ private:
     // SuperVolo timer for min RPM and very basic synthetic air speed
     uint32_t synthetic_arspd_ms;
     uint32_t synthetic_arspd_message_ms;
-    int8_t last_synthetic_arspd;
+    int16_t last_synthetic_arspd;
 };
